add css 🤦‍♀️🤦‍♀️🤦‍♀️🤦‍♀️

+++ src/components/Posts.js
@@ -3,6 +3,8 @@ import { useSelector, useDispatch } from "react-redux";
 import * as actions from "../actions/posts";
 import * as functions_Handeler from "../actions/API_Funtions_Handeler"
 import { Form, Modal, Button, FormGroup, FormControl } from 'react-bootstrap';
+import Swal from 'sweetalert2';
+
 
 
 export default function Posts() {
@@ -12,39 +14,39 @@ export default function Posts() {
     const dispatch = useDispatch();
 
 
-
-    // useEffect(() => {
-    //     async function fetchData() {
-    //         if (posts.length === 0) {
-    //             const posts_Data = await functions_Handeler.getPosts();
-
-    //             dispatch(actions.getPosts(posts_Data));
-    //         }
-
-    //     }
-    //     fetchData();
-    // }, [])
-
-
-    //function delete posts 
-
     const deletePost = async (idx) => {
-        // console.log(` posts before delete ${posts}`)
-        let delpost = await functions_Handeler.deletePostFun(posts, idx);
-        dispatch(actions.deletePosts(delpost))
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then(async(result) => {
+            if (result.isConfirmed) {
+
+                let newPosts = await functions_Handeler.deletePostFun(posts, idx);
+                dispatch(actions.deletePosts(newPosts));
+
+                Swal.fire(
+                    'Deleted!',
+                    'Your post has been deleted.',
+                    'success'
+                )
+            }
+        })
+
 
     }
 
-    //use state to show model
     const [showUpdate, setShowUpdate] = useState(false);
     const [updatePost, setUpdatePost] = useState({});
-    //showUpdate function 
+
     const showUpdateForm = (myPost) => {
         setShowUpdate(!showUpdate);
         setUpdatePost(myPost);
     }
-
-    //update function 
 
     const updatePostFun = async (e) => {
         e.preventDefault();
@@ -63,8 +65,6 @@ export default function Posts() {
         showUpdateForm();
     }
 
-    //function to add new post 
-
     const addPostHandler = async (e) => {
         e.preventDefault();
         let newPost = {
@@ -80,52 +80,14 @@ export default function Posts() {
         dispatch(actions.addPosts(newPost));
     }
 
-
     return (
-        <div  className=' flex flex-col justify-center items-center;'>
-            <h3  className=' flex flex-col justify-center items-center;'> Add New Post </h3>
-            <div >
-            <form class="w-full max-w-sm  " onSubmit={addPostHandler}>
-  <div class="md:flex md:items-center mb-6">
-    <div class="md:w-1/3">
-      <label class="block text-gray-500 font-bold md:text-right mb-1 md:mb-0 pr-4" for="inline-full-name">
-      Title
-      </label>
-    </div>
-    <div class="md:w-2/3">
-      <textarea   name="title" id="title"   as="textarea" class="bg-gray-200 appearance-none border-2 border-gray-200 rounded w-full py-2 px-4 text-gray-700 leading-tight focus:outline-none focus:bg-white focus:border-purple-500"  placeholder='Title ...' type="text"/>
-    </div>
-  </div>
-  <div class="md:flex md:items-center mb-6">
-    <div class="md:w-1/3">
-      <label class="block text-gray-500 font-bold md:text-right mb-1 md:mb-0 pr-4" for="inline-password">
-      Body
-      </label>
-    </div>
-    <div class="md:w-2/3">
-      <textarea  name="body"  as="textarea" id="body" class="bg-gray-200 appearance-none border-2 border-gray-200 rounded w-full py-2 px-4 text-gray-700 leading-tight focus:outline-none focus:bg-white focus:border-purple-500" type="text"/>
-    </div>
-  </div>
-  <div class="md:flex md:items-center mb-6">
-    <div class="md:w-1/3"></div>
+        <div>
+            <h1 className="text-center">Posts</h1>
 
-  </div>
-  <div class="md:flex md:items-center">
-    <div class="md:w-1/3"></div>
-    <div class="md:w-2/3">
-    <input type='submit' value="Add" />
-      
-    </div>
-  </div>
-</form>
-</div>
             <div>
+                <h3> Add New Post </h3>
 
-
-
-
-
-                {/* <Form onSubmit={addPostHandler}>
+                <Form onSubmit={addPostHandler}>
                     <FormGroup>
                         <Form.Label>Title: </Form.Label>
                         <Form.Control as="textarea" placeholder='Title ...' name="title" type="text" />
@@ -140,60 +102,24 @@ export default function Posts() {
                         Add
                     </Button>
                 </Form>
-
- */}
-
-
-
-
-
-
-
-
-
-
-                {/* <div class="max-w-sm rounded overflow-hidden shadow-lg">
-
-                    <div class="px-6 py-4">
-                        <div class="font-bold text-xl mb-2">{myPost.body}</div>
-                        <p class="text-gray-700 text-base">
-                            {myPost.body}
-                        </p>
-                    </div> */}
-                {/* <div class="px-6 pt-4 pb-2">
-    <span class="inline-block bg-gray-200 rounded-full px-3 py-1 text-sm font-semibold text-gray-700 mr-2 mb-2">#photography</span>
-    <span class="inline-block bg-gray-200 rounded-full px-3 py-1 text-sm font-semibold text-gray-700 mr-2 mb-2">#travel</span>
-    <span class="inline-block bg-gray-200 rounded-full px-3 py-1 text-sm font-semibold text-gray-700 mr-2 mb-2">#winter</span>
-  </div> */}
-                {/* </div> */}
-
-
-
-
-
-
-
-
-
-
             </div>
-            <div class="  grid grid-cols-3 gap-4 place-items-start  ..." >
-                {
+            <div class="grid grid-cols-3 gap-4 place-items-start  ..." >
+            {
                     posts.map((myPost, idx) => {
                         return (
                             <div key={idx}>
                                 {/* <h2>{myPost.title}</h2>
                             <p>{myPost.body}</p> */}
-                                <div class=" h-120 w-100  max-w-sm rounded overflow-hidden shadow-lg ">
+                                <div class=" w-100  max-w-sm rounded overflow-hidden shadow-lg ">
 
-                                    <div class="px-6 py-4 " >
+                                    <div class="px-3 py-2 " >
                                         <div class="font-bold text-xl mb-2">{myPost.title}</div>
                                         <p class="  flex flex-nowrap text-gray-700 text-base">
                                             {myPost.body}
                                         </p>
                                     </div>
 
-                                    <ul>
+                                    <ul className="px-3">
                                         {
                                             myPost.comments.map((comm, index) => {
                                                 return (
@@ -205,9 +131,9 @@ export default function Posts() {
 
 
 
-                                                        <div class="px-6 pt-4 pb-2">
-                                                            <span class="inline-block bg-gray-200 rounded-full px-3 py-1 text-sm font-semibold text-gray-700 mr-2 mb-2">{comm.name}</span>
-                                                            <span class="inline-block bg-gray-200 rounded-full px-3 py-1 text-sm font-semibold text-gray-700 mr-2 mb-2">{comm.body}</span></div>
+                                                        <div class="border-b-2 border-blue-600 bg-gray-200  px-3 pt-4 pb-2">
+                                                            <span class="font-black inline-block  px-3 py-1 text-sm font-semibold text-gray-700 mr-2 mb-2">{comm.name}</span>
+                                                            <span class="inline-block  px-3 py-1 text-sm font-semibold text-gray-700 mr-2 mb-2">{comm.body}</span></div>
                                                     </li>
 
                                                 )
@@ -236,7 +162,36 @@ export default function Posts() {
 
 
                 }
-            </div>
+                </div>
+            {/* {
+                posts.map((myPost, idx) => {
+                    return (
+                        <div key={idx}>
+                            {
+                                myPost.userId == myInfos.id &&
+                                <>
+                                    <button onClick={() => deletePost(myPost.id)}> Delete </button>
+                                    <button onClick={() => showUpdateForm(myPost)}> Update </button>
+                                </>
+                            }
+                            <h2>{myPost.title}</h2>
+                            <p>{myPost.body}</p>
+                            <ul>
+                                {
+                                    myPost.comments.map((comm, index) => {
+                                        return (
+                                            <li key={index}>
+                                                <h3>{comm.name}</h3>
+                                                <p>{comm.body}</p>
+                                            </li>
+                                        )
+                                    })
+                                }
+                            </ul>
+                        </div>
+                    )
+                })
+            } */}
             <Modal show={showUpdate} onHide={showUpdateForm}>
                 <Modal.Header closeButton>
                     <Modal.Title>Update Post</Modal.Title>
@@ -263,7 +218,6 @@ export default function Posts() {
 
 
             </Modal>
-
         </div>
     )
 }
